+++ src/main/kotlin/com/github/israelermel/iridio77/impl/IdeaNotificationsManagerImplementation.kt
@@ -1,0 +1,5 @@
+package com.github.israelermel.iridio77.impl
+
+interface IdeaNotificationsManagerImplementation {
+    fun showNotification(title: String, message: String)
+}