+++ src/main/kotlin/com/github/israelermel/iridio77/models/AndroidDebugEvent.kt
@@ -1,0 +1,9 @@
+package com.github.israelermel.iridio77.models
+
+enum class AndroidDebugEvent {
+    SHOW_LAYOUT_BOUNDS,
+    TOOGLE_TALKBACK,
+    TOOGLE_ANIMATIONS,
+    TOOGLE_PROFILE,
+    TOOGLE_OVERDRAW
+}