+++ src/main/kotlin/com/github/israelermel/iridio77/ui/models/Command.kt
@@ -1,0 +1,6 @@
+package com.github.israelermel.iridio77.ui.models
+
+interface Command {
+    fun getCommand(): String
+    fun getLabel(): String
+}