+++ CHANGELOG.md
@@ -3,5 +3,16 @@
 # iridio77 Changelog
 
 ## [Unreleased]
+
+## [0.0.4] - 2023-01-18
+
 ### Added
-- Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+- This is the first version containing commands to change Emulator/Physical Device settings such as:
+- Activate and deactivate TalkBack
+- Change Density
+- Change Font Size
+- Activate and deactivate Layout Bounds, Profile, Overdraw and Animations
+- Add custom attributes in local.properties via Preferences
+
+[Unreleased]: https://github.com/israelermel/iridio77/compare/v0.0.4...HEAD
+[0.0.4]: https://github.com/israelermel/iridio77/commits/v0.0.4